refactor into trace.go
