+++ pkg/evm/evm.go
@@ -22,7 +22,6 @@ import (
 	"github.com/ava-labs/subnet-evm/params"
 	"github.com/ava-labs/subnet-evm/precompile/contracts/warp"
 	"github.com/ava-labs/subnet-evm/predicate"
-	"github.com/ava-labs/subnet-evm/rpc"
 	subnetEvmUtils "github.com/ava-labs/subnet-evm/utils"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
@@ -35,8 +34,6 @@ const (
 	maxPriorityFeePerGas        = 2500000000 // 2.5 gwei
 	nativeTransferGas    uint64 = 21_000
 )
-
-var ErrUnknownErrorSelector = fmt.Errorf("unknown error selector")
 
 // wraps over ethclient for calls used by SDK. includes:
 // - finds out url scheme in case it is missing, to connect to ws/wss/http/https
@@ -468,15 +465,17 @@ func (client Client) WaitForEVMBootstrapped(timeout time.Duration) error {
 	}
 }
 
-func GetTxToMethodWithWarpMessage(
-	client Client,
-	generateRawTxOnly bool,
+// Generates a transaction signed with [privateKeyStr], calling a [contract] method using [callData]
+// including [warpMessage] in the tx accesslist
+// If [generateRawTxOnly] is set, it generates a similar, unsigned tx, with given [from] address
+func (client Client) TransactWithWarpMessage(
 	from common.Address,
 	privateKeyStr string,
 	warpMessage *avalancheWarp.Message,
 	contract common.Address,
 	callData []byte,
 	value *big.Int,
+	generateRawTxOnly bool,
 ) (*types.Transaction, error) {
 	const defaultGasLimit = 2_000_000
 	var (
@@ -561,106 +560,6 @@ func GetEventFromLogs[T any](logs []*types.Log, parser func(log types.Log) (T, e
 		cumErrMsg += fmt.Sprintf("log %d -> %s", i, err.Error())
 	}
 	return *new(T), fmt.Errorf("failed to find %T event in receipt logs: [%s]", *new(T), cumErrMsg)
-}
-
-func GetRPCClient(rpcURL string) (*rpc.Client, error) {
-	var (
-		client *rpc.Client
-		err    error
-	)
-	hasScheme, err := HasScheme(rpcURL)
-	if err != nil {
-		return nil, err
-	}
-	for i := 0; i < repeatsOnFailure; i++ {
-		ctx, cancel := utils.GetAPILargeContext()
-		defer cancel()
-		if !hasScheme {
-			_, scheme, findErr := GetClientWithoutScheme(rpcURL)
-			if findErr == nil {
-				client, err = rpc.DialContext(ctx, scheme+rpcURL)
-			} else {
-				err = findErr
-			}
-		} else {
-			client, err = rpc.DialContext(ctx, rpcURL)
-		}
-		if err == nil {
-			break
-		}
-		err = fmt.Errorf("failure connecting to rpc client on %s: %w", rpcURL, err)
-		ux.Logger.RedXToUser("%s", err)
-		time.Sleep(sleepBetweenRepeats)
-	}
-	return client, err
-}
-
-func DebugTraceTransaction(
-	client *rpc.Client,
-	txID string,
-) (map[string]interface{}, error) {
-	var (
-		err   error
-		trace map[string]interface{}
-	)
-	for i := 0; i < repeatsOnFailure; i++ {
-		ctx, cancel := utils.GetAPILargeContext()
-		defer cancel()
-		err = client.CallContext(
-			ctx,
-			&trace,
-			"debug_traceTransaction",
-			txID,
-			map[string]string{"tracer": "callTracer"},
-		)
-		if err == nil {
-			break
-		}
-		err = fmt.Errorf("failure tracing tx %s for client %#v: %w", txID, client, err)
-		time.Sleep(sleepBetweenRepeats)
-	}
-	return trace, err
-}
-
-func DebugTraceCall(
-	client *rpc.Client,
-	toTrace map[string]string,
-) (map[string]interface{}, error) {
-	var (
-		err   error
-		trace map[string]interface{}
-	)
-	for i := 0; i < repeatsOnFailure; i++ {
-		ctx, cancel := utils.GetAPILargeContext()
-		defer cancel()
-		err = client.CallContext(
-			ctx,
-			&trace,
-			"debug_traceCall",
-			toTrace,
-			"latest",
-			map[string]interface{}{
-				"tracer": "callTracer",
-				"tracerConfig": map[string]interface{}{
-					"onlyTopCall": false,
-				},
-			},
-		)
-		if err == nil {
-			break
-		}
-		err = fmt.Errorf("failure tracing call for client %#v: %w", client, err)
-		time.Sleep(sleepBetweenRepeats)
-	}
-	return trace, err
-}
-
-func GetTrace(rpcURL string, txID string) (map[string]interface{}, error) {
-	client, err := GetRPCClient(rpcURL)
-	if err != nil {
-		return nil, err
-	}
-	return DebugTraceTransaction(client, txID)
 }
 
 func SetupProposerVM(
@@ -777,39 +676,6 @@ func ExtractWarpMessageFromReceipt(
 	}
 	txLog := logs[0]
 	return warp.UnpackSendWarpEventDataToMessage(txLog.Data)
-}
-
-func GetFunctionSelector(functionSignature string) string {
-	return "0x" + hex.EncodeToString(crypto.Keccak256([]byte(functionSignature))[:4])
-}
-
-func GetErrorFromTrace(
-	trace map[string]interface{},
-	functionSignatureToError map[string]error,
-) (error, error) {
-	traceOutputI, ok := trace["output"]
-	if !ok {
-		return nil, fmt.Errorf("trace does not contain output field")
-	}
-	traceOutput, ok := traceOutputI.(string)
-	if !ok {
-		return nil, fmt.Errorf("expected type string for trace output, got %T", traceOutputI)
-	}
-	traceOutputBytes, err := hex.DecodeString(strings.TrimPrefix(traceOutput, "0x"))
-	if err != nil {
-		return nil, fmt.Errorf("failure decoding trace output: %w", err)
-	}
-	if len(traceOutputBytes) < 4 {
-		return nil, fmt.Errorf("less than 4 bytes in trace output")
-	}
-	traceErrorSelector := "0x" + hex.EncodeToString(traceOutputBytes[:4])
-	for errorSignature, err := range functionSignatureToError {
-		errorSelector := GetFunctionSelector(errorSignature)
-		if traceErrorSelector == errorSelector {
-			return err, nil
-		}
-	}
-	return nil, fmt.Errorf("%w: %s", ErrUnknownErrorSelector, traceErrorSelector)
 }
 
 func TransactionError(tx *types.Transaction, err error, msg string, args ...interface{}) error {
